COVID-19 Data Analysis and Vaccination Impact Study
This project aims to analyze the global COVID-19 pandemic data, focusing on the spread of the virus, the impact on different countries and continents, and the effectiveness of vaccination campaigns. The project involves data exploration, cleansing, and analysis techniques, including SQL queries to extract, clean, and analyze data from the CovidProject database. Techniques such as data partitioning, window functions, and temporary tables are used to derive meaningful insights from the data. The project also includes the calculation of various metrics such as infection rates, death percentages, and vaccination percentages to understand the impact of COVID-19 and vaccination efforts on different regions.